refactor: :recycle: Réorganiser la structure des fichiers et des imports
