+++ src/App.tsx
@@ -1,12 +1,11 @@
 import { Outlet } from 'react-router-dom';
 import { useEffect } from 'react';
-import './App.css';
-import Footer from '../Footer/Footer';
+import Footer from './components/Footer/Footer';
 
 // import { CartProvider } from '../Cart/CartContext/CartContext';
-import { useUser } from '../../context/UserContext';
-import api from '../../api';
-import Header from '../Header/Header';
+import { useUser } from './context/UserContext';
+import api from './api';
+import Header from './components/Header/Header';
 
 function App() {
   const { setUser } = useUser();
