+++ src/main.tsx
@@ -4,7 +4,7 @@ import router from './router';
 import './styles/index.css';
 import { UserProvider } from './context/UserContext';
 import { ProjectProvider } from './context/ProjectContext';
-import { CartProvider } from './components/Cart/CartContext/CartContext';
+import { CartProvider } from './pages/Cart/CartContext/CartContext';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   // <React.StrictMode>
