+++ src/pages/HomePage/TreesHighlights.tsx
