+++ src/pages/LegalNotices/LegalNotices.tsx
