+++ src/router.tsx
@@ -1,23 +1,23 @@
 import { createBrowserRouter } from 'react-router-dom';
 import ProtectedRoute from './auth/ProtectedRoute';
-import App from './components/App/App';
-import Register from './components/Register/Register';
-import HomePage from './components/HomePage/HomePage';
-import LoginPage from './components/LoginPage/LoginPage';
-import DetailTreePage from './components/DetailTreePage/DetaiTreePage';
-import ProjectDetails from './components/ProjectDetailsPage/ProjectDetails';
-import ProjectPage from './components/ProjectPage/ProjectPage';
-import Contact from './components/Contact/Contact';
-import IntroduceGr from './components/IntroducieGr/IntroduceGr';
-import LegalNotices from './components/LegalNotices/LegalNotices';
-import Cart from './components/Cart/Cart';
-import OrdersPage from './components/OrdersPage/OrdersPage';
-import OrderDetailPage from './components/OrderDetailPage/OrderDetailPage';
-import ProfilDetails from './components/ProfilDetails/ProfilDetails';
-import MyAccountPage from './components/MyAccountPage/MyAccountPage';
-import Stripe from './components/Stripe/Stripe';
-import ConfirmPay from './components/Stripe/ConfirmPay';
-import EditProfil from './components/EditProfil/EditProfil';
+import App from './App';
+import Register from './pages/Register/Register';
+import HomePage from './pages/HomePage/HomePage';
+import LoginPage from './pages/LoginPage/LoginPage';
+import DetailTreePage from './pages/DetailTreePage/DetaiTreePage';
+import ProjectDetails from './pages/ProjectDetailsPage/ProjectDetails';
+import ProjectPage from './pages/ProjectPage/ProjectPage';
+import Contact from './pages/Contact/Contact';
+import IntroduceGr from './pages/IntroducieGr/IntroduceGr';
+import LegalNotices from './pages/LegalNotices/LegalNotices';
+import Cart from './pages/Cart/Cart';
+import OrdersPage from './pages/OrdersPage/OrdersPage';
+import OrderDetailPage from './pages/OrderDetailPage/OrderDetailPage';
+import ProfilDetails from './pages/ProfilDetails/ProfilDetails';
+import MyAccountPage from './pages/MyAccountPage/MyAccountPage';
+import Stripe from './pages/Stripe/Stripe';
+import ConfirmPay from './pages/Stripe/ConfirmPay';
+import EditProfil from './pages/EditProfil/EditProfil';
 
 const router = createBrowserRouter([
   {
